es6 for client

+++ client/libraries/essentials/addToStorage.js
@@ -1,9 +1,11 @@
-function addToStorage(data = window._RUNTIME_DATA_) {
+function addToStorage(data = window._RUNTIME_DATA_, isForced) {
   const dataKeys = Object.keys(data)
 
   let _len = dataKeys.length
   while (_len--) {
     const key = dataKeys[_len]
-    sessionStorage.setItem(key, JSON.stringify(data[key]))
+
+    if (!sessionStorage.getItem(key) || isForced)
+      sessionStorage.setItem(key, JSON.stringify(data[key]))
   }
 }
